Update deps
The eslint-plugin-functional pin is intentional, it consistenly fails to parse src/String.ts for some reason.

The docs changes are due to breaking changes in docs-ts, but all these changes are desirable.

The new @types/jest dependency is now needed, so that's probably a breaking change on Jest's part that doesn't adhere to semver.